requirements.txt
# requirements.txt
Flask>=2.0
kafka-python>=2.0
psycopg2-binary>=2.9
